Clean up naming and remove "optimized" terminology in project
