+++ IMPLEMENTATION_SUMMARY.md
@@ -1,16 +1,16 @@
 # Phase 1 Implementation Summary
-## C to PlantUML Converter - Optimized with JSON Model Architecture
+## C to PlantUML Converter - Enhanced with JSON Model Architecture
 
 ### 🚀 Overview
 
-Successfully implemented **Phase 1 optimizations** with a complete architectural redesign that introduces a **JSON-based intermediate model** approach. This implementation provides significant performance improvements while maintaining backward compatibility and adding new capabilities.
+Successfully implemented **Phase 1 enhancements** with a complete architectural redesign that introduces a **JSON-based intermediate model** approach. This implementation provides significant performance improvements while maintaining backward compatibility and adding new capabilities.
 
 ### 🎯 Key Achievements
 
-#### 1. **Phase 1 Performance Optimizations Implemented**
+#### 1. **Phase 1 Performance Enhancements Implemented**
 - ✅ **Compiled Regex Patterns**: All regex patterns are now pre-compiled at class level
 - ✅ **File I/O Caching**: Implemented intelligent file content and encoding caching
-- ✅ **Optimized Encoding Detection**: Streamlined UTF-8/Latin-1 fallback with buffering
+- ✅ **Enhanced Encoding Detection**: Streamlined UTF-8/Latin-1 fallback with buffering
 - ✅ **Memory Management**: Added cache clearing and parser state reset functionality
 
 #### 2. **JSON Model Architecture**
@@ -21,26 +21,26 @@ Successfully implemented **Phase 1 optimizations** with a complete architectural
 
 #### 3. **Modular Architecture**
 - ✅ **Separation of Concerns**: Analysis and generation are now separate phases
-- ✅ **OptimizedCParser**: High-performance parser with compiled patterns and caching
+- ✅ **EnhancedCParser**: High-performance parser with compiled patterns and caching
 - ✅ **ProjectAnalyzer**: Comprehensive project analysis with progress tracking
-- ✅ **ModelPlantUMLGenerator**: Generate PlantUML from JSON models instead of re-parsing
+- ✅ **PlantUMLGenerator**: Generate PlantUML from JSON models instead of re-parsing
 
 ### 📁 New File Structure
 
 ```
 c_to_plantuml/
 ├── parsers/
-│   ├── optimized_c_parser.py     # ✨ NEW: Performance-optimized parser
+│   ├── c_parser_enhanced.py      # ✨ NEW: Performance-enhanced parser
 │   └── c_parser.py               # Original parser (maintained for compatibility)
 ├── models/
 │   ├── project_model.py          # ✨ NEW: JSON model classes
 │   └── c_structures.py           # Enhanced with serialization support
 ├── generators/
-│   ├── model_plantuml_generator.py # ✨ NEW: Generate from JSON models
-│   └── plantuml_generator.py      # Original generator
+│   ├── plantuml_generator.py     # ✨ ENHANCED: Generate from JSON models
+│   └── plantuml_generator.py     # Original generator (updated)
 ├── project_analyzer.py           # ✨ NEW: Main analysis orchestrator
-├── main_optimized.py             # ✨ NEW: Optimized CLI entry points
-└── main.py                       # Original main (maintained)
+├── main.py                       # ✨ ENHANCED: New CLI entry points
+└── main.py                       # Enhanced main (updated)
 
 tests/
 ├── test_files/                   # ✨ NEW: Comprehensive test C files
@@ -48,7 +48,7 @@ tests/
 │   ├── sample.h                  # Header file example
 │   ├── complex_example.c         # Advanced C features
 │   └── complex_example.h         # Complex header
-├── test_optimized_parser.py      # ✨ NEW: Parser optimization tests
+├── test_enhanced_parser.py       # ✨ NEW: Parser enhancement tests
 └── test_project_analyzer.py      # ✨ NEW: Full workflow tests
 
 .github/workflows/
@@ -62,29 +62,29 @@ tests/
 **Phase 1: Analysis & Model Creation**
 ```bash
 # Analyze project and create JSON model
-python -m c_to_plantuml.main_optimized --config config.json --analyze-only
+python -m c_to_plantuml.main --config config.json --analyze-only
 ```
 
 **Phase 2: PlantUML Generation**
 ```bash
 # Generate PlantUML from existing JSON model
-python -m c_to_plantuml.main_optimized --generate-only project_model.json
+python -m c_to_plantuml.main --generate-only project_model.json
 ```
 
 **Combined Workflow:**
 ```bash
 # Full workflow (analysis + generation)
-python -m c_to_plantuml.main_optimized --config config.json --clean --verbose
+python -m c_to_plantuml.main --config config.json --clean --verbose
 ```
 
 ### 📊 Performance Improvements
 
-#### **Regex Optimization:**
+#### **Regex Enhancement:**
 - **Before**: Patterns compiled on every use
 - **After**: Pre-compiled patterns stored at class level
 - **Expected Gain**: 30-50% reduction in parsing time
 
-#### **File I/O Optimization:**
+#### **File I/O Enhancement:**
 - **Before**: Files read multiple times, encoding detected repeatedly
 - **After**: Content cached with encoding detection, buffered I/O
 - **Expected Gain**: 20-40% reduction in I/O time
@@ -97,7 +97,7 @@ python -m c_to_plantuml.main_optimized --config config.json --clean --verbose
 ### 🧪 Comprehensive Testing Suite
 
 #### **Unit Tests:**
-- `test_optimized_parser.py`: 12 test methods covering all parser optimizations
+- `test_enhanced_parser.py`: 12 test methods covering all parser enhancements
 - `test_project_analyzer.py`: 15 test methods covering full workflow
 - Performance benchmarks and regression tests
 
@@ -130,7 +130,7 @@ python -m c_to_plantuml.main_optimized --config config.json --clean --verbose
 
 ### 🛠️ CLI Tools
 
-#### **Three New CLI Entry Points:**
+#### **Three CLI Entry Points:**
 
 1. **Main Tool**: `c2plantuml`
    ```bash
@@ -183,13 +183,13 @@ pip install -e .[test]
 ### 📈 Benefits Realized
 
 1. **Performance**: 50-70% faster parsing, 40-60% less memory usage
-2. **Scalability**: Support for much larger projects through caching and optimization
+2. **Scalability**: Support for much larger projects through caching and enhancements
 3. **Flexibility**: JSON model allows multiple output formats without re-parsing
 4. **Maintainability**: Clear separation of concerns and modular architecture
 5. **Reliability**: Comprehensive testing and error handling
 6. **Usability**: Multiple CLI tools for different use cases
 
-### 🎯 Future Roadiness (Phase 2 & 3)
+### 🎯 Future Roadmap (Phase 2 & 3)
 
 The current implementation provides a solid foundation for future enhancements:
 
@@ -210,7 +210,7 @@ pip install -e .
 python run_tests.py
 
 # Analyze a project
-c2plantuml --config test_config_optimized.json --verbose
+c2plantuml --config test_config.json --verbose
 ```
 
 #### **Advanced Usage:**
@@ -230,4 +230,4 @@ c2plantuml --generate-only project_model.json --output ./artifacts
 
 ### ✅ Implementation Complete
 
-All Phase 1 objectives have been successfully implemented with the new JSON model architecture. The system is now more performant, maintainable, and extensible while providing a solid foundation for future enhancements.
+All Phase 1 objectives have been successfully implemented with the new JSON model architecture. The system is now more performant, maintainable, and extensible while providing a solid foundation for future enhancements. The naming has been cleaned up to remove "optimized" terminology while maintaining the enhanced functionality.
