+++ test_config.json
@@ -1,13 +1,15 @@
 {
+  "project_name": "Test_C_Project",
   "project_roots": [
-    "D:/WORK/Sandbox/doom/ipx", "D:/WORK/Sandbox/doom/linuxdoom-1.10", "D:/WORK/Sandbox/doom/sersrc","D:/WORK/Sandbox/doom/sndserv"
+    "./tests/test_files"
   ],
   "output_dir": "./output_uml",
   "output_dir_packaged": "./output_packaged",
+  "model_output_path": "./project_model.json",
   "recursive": true,
   "c_file_prefixes": [],
   "structure_overrides": [
-    {"from_puml": "am_map.puml", "to_folder": "special_folder"},
-    {"from_puml": "vehicle.puml", "to_folder": "vehicles"}
+    {"from_puml": "sample.puml", "to_folder": "core"},
+    {"from_puml": "complex_example.puml", "to_folder": "advanced"}
   ]
 }
